[docker] move dockeragentparams up beside agentparams
