+++ scenarios/aws/dockerVM/run.go
@@ -2,7 +2,7 @@ package dockervm
 
 import (
 	"github.com/DataDog/test-infra-definitions/components/datadog/agent/docker"
-	"github.com/DataDog/test-infra-definitions/components/datadog/agent/dockerparams"
+	"github.com/DataDog/test-infra-definitions/components/datadog/dockeragentparams"
 	resourcesAws "github.com/DataDog/test-infra-definitions/resources/aws"
 	"github.com/DataDog/test-infra-definitions/scenarios/aws/utils"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
@@ -14,9 +14,9 @@ func Run(ctx *pulumi.Context) error {
 		return err
 	}
 
-	var options []dockerparams.Option
+	var options []dockeragentparams.Option
 	if env.AgentDeploy() {
-		options = append(options, dockerparams.WithAgent())
+		options = append(options, dockeragentparams.WithAgent())
 	}
 
 	architecture, err := utils.GetArchitecture(env.GetCommonEnvironment())
@@ -24,7 +24,7 @@ func Run(ctx *pulumi.Context) error {
 		return err
 	}
 
-	options = append(options, dockerparams.WithArchitecture(architecture))
+	options = append(options, dockeragentparams.WithArchitecture(architecture))
 
 	_, err = docker.NewDaemonWithEnv(env, options...)
